アクセスできるファイルに権限を付与、error時はerrorHandler.tsでerror.tsxに遷移

+++ app/components/Hooks/useMethod.ts
@@ -1,30 +1,44 @@
-import { getKey } from "../../store/auth/keySlice";
+import { getKey, KeyState } from "../../store/auth/keySlice";
 import { isLogout } from "../../store/auth/isLoginSlice";
 import { logout } from "../../store/auth/userSlice";
+import { AppDispatch } from "../../redux/store";
+import { NextRouter } from "next/router";
 
-export const allLogout = async (dispatch: any): Promise<string | null> => {
+export const allLogout = async (
+  dispatch: AppDispatch
+): Promise<string | null> => {
   try {
-    await dispatch(logout({}) as any);
-    dispatch(isLogout());
-    localStorage.clear();
+    const response = await dispatch(logout() as any);
+    if (response.meta.requestStatus === "fulfilled") {
+      dispatch(isLogout());
+      localStorage.clear();
+      return "ログアウトしました";
+    }
+    return null;
   } catch (error) {
     console.log("Error", error);
     return null;
   }
 };
 
-export const getUserKey = async (dispatch: any): Promise<string | null> => {
+export const getUserKey = async (
+  dispatch: AppDispatch
+): Promise<KeyState | null> => {
   try {
-    const response = await dispatch(getKey({}) as any);
-    console.log("responseGetKey", response);
-    return response.payload.roleKey;
+    const response = await dispatch(getKey() as any);
+    if (response.meta.requestStatus === "fulfilled") {
+      console.log("responseGetKey", response);
+
+      return response.payload.roleKey;
+    }
+    return null;
   } catch (error) {
     console.log("Error", error);
     return null;
   }
 };
 
-export const ownerPermission = (role: string, router: any): boolean => {
+export const ownerPermission = (role: string, router: NextRouter): boolean => {
   if (role === "オーナー") {
     return true;
   } else {
@@ -32,7 +46,10 @@ export const ownerPermission = (role: string, router: any): boolean => {
   }
 };
 
-export const managerPermission = (role: string, router: any): boolean => {
+export const managerPermission = (
+  role: string,
+  router: NextRouter
+): boolean => {
   if (role === "マネージャー" || role === "オーナー") {
     return true;
   } else {
@@ -40,7 +57,7 @@ export const managerPermission = (role: string, router: any): boolean => {
   }
 };
 
-export const staffPermission = (role: string, router: any): boolean => {
+export const staffPermission = (role: string, router: NextRouter): boolean => {
   if (role === "スタッフ" || role === "マネージャー" || role === "オーナー") {
     return true;
   } else {
